Fixed some issues on synthesis due to compiler behavior, added copyright notice and thanks to readme for TinySoundFont and SFZero.

+++ Source/AlphaTab.CSharp/Platform/CSharp/AlphaSynthWorkerApiBase.cs
@@ -140,7 +140,7 @@ namespace AlphaTab.Platform.CSharp
             DispatchOnWorkerThread(() => { Player.SetChannelSolo(channel, solo); });
         }
 
-        public void SetChannelVolume(int channel, double volume)
+        public void SetChannelVolume(int channel, float volume)
         {
             DispatchOnWorkerThread(() => { Player.SetChannelVolume(channel, volume); });
         }
